+++ common/src/main/java/com/telepathicgrunt/the_bumblezone/blocks/HeavyAir.java
@@ -9,6 +9,7 @@ import com.telepathicgrunt.the_bumblezone.modinit.BzBlocks;
 import com.telepathicgrunt.the_bumblezone.modinit.BzItems;
 import com.telepathicgrunt.the_bumblezone.modinit.BzParticles;
 import com.telepathicgrunt.the_bumblezone.modinit.BzTags;
+import com.telepathicgrunt.the_bumblezone.utils.PlatformHooks;
 import net.minecraft.core.BlockPos;
 import net.minecraft.core.Holder;
 import net.minecraft.core.HolderSet;
@@ -129,9 +130,7 @@ public class HeavyAir extends Block {
             }
 
             if (player.getAbilities().flying) {
-                player.getAbilities().flying = false;
-                player.getAbilities().mayfly = false;
-                player.onUpdateAbilities();
+                PlatformHooks.disableFlight(player);
             }
 
             if (entity instanceof TemporaryPlayerData temporaryPlayerData) {